requirements.txt
colorama>=0.4.6
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
fpdf>=1.7.2
gunicorn>=23.0.0
openpyxl>=3.1.5
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
rich>=14.0.0
supabase 